Regression related to automatic logout and guest account for issue #3343

+++ lib/auth.php
@@ -713,26 +713,29 @@ function is_realm_allowed($realm) {
 		}
 
 		if (!is_user_perms_valid($_SESSION['sess_user_id'])) {
-			kill_session_var('sess_user_realms');
-			kill_session_var('sess_user_config_array');
-			kill_session_var('sess_config_array');
-
 			if (db_table_exists('user_auth_cache')) {
 				$enabled = db_fetch_cell_prepared('SELECT enabled
 					FROM user_auth
 					WHERE id = ?',
 					array($_SESSION['sess_user_id']));
 
-				if ($enabled == '') {
+				if ($enabled == '' && get_guest_account() != $_SESSION['sess_user_id']) {
 					db_execute_prepared('DELETE FROM user_auth_cache
 						WHERE user_id = ?',
 						array($_SESSION['sess_user_id']));
 
 					kill_session_var('sess_user_id');
+					kill_session_var('sess_user_realms');
+					kill_session_var('sess_user_config_array');
+					kill_session_var('sess_config_array');
 
 					print '<span style="display:none;">cactiLoginSuspend</span>';
 					exit;
 				}
+			} else {
+				kill_session_var('sess_user_realms');
+				kill_session_var('sess_user_config_array');
+				kill_session_var('sess_config_array');
 			}
 		}
